Module ocicrypt_rs::blockcipher

Structs

LayerBlockCipherOptions

The public and private configuration information required to encrypt/decrypt an image layer.

PrivateLayerBlockCipherOptions

The information required to encrypt/decrypt an image layer which are sensitive and should not be in plaintext.

PublicLayerBlockCipherOptions

The information required to encrypt/decrypt an image layer which are public and can be deduplicated in plaintext across multiple recipients.

Enums

LayerBlockCipherHandler

Handler for image layer encryption/decryption.

Constants

AES256CTR

The default cipher algorithm for image layer encryption/decryption.

Traits

EncryptionFinalizer

Trait to finalize the encryption operation.

LayerBlockCipher

Trait to setup the encryption/decryption context.

Type Aliases

LayerCipherType

Type of the cipher algorithm used to encrypt/decrypt image layers.