Trait ocicrypt_rs::keywrap::KeyWrapper

pub trait KeyWrapper: Send + Sync {
    // Required methods
    fn wrap_keys(&self, ec: &EncryptConfig, opts_data: &[u8]) -> Result<Vec<u8>>;
    fn unwrap_keys(
        &self,
        dc: &DecryptConfig,
        annotation: &[u8]
    ) -> Result<Vec<u8>>;
    fn annotation_id(&self) -> String;
    fn probe(&self, dc_param: &HashMap<String, Vec<Vec<u8>>>) -> bool;

    // Provided methods
    fn private_keys(
        &self,
        dc_param: &HashMap<String, Vec<Vec<u8>>>
    ) -> Option<Vec<Vec<u8>>> { ... }
    fn keyids_from_packet(&self, packet: String) -> Option<Vec<u64>> { ... }
    fn recipients(&self, recipients: String) -> Option<Vec<String>> { ... }
}

KeyWrapper is the interface used for wrapping keys using a specific encryption technology (pgp, jwe, pkcs7, pkcs11, keyprovider)

Required Methods

fn wrap_keys(&self, ec: &EncryptConfig, opts_data: &[u8]) -> Result<Vec<u8>>

wrap keys data with encrypt config.

fn unwrap_keys(&self, dc: &DecryptConfig, annotation: &[u8]) -> Result<Vec<u8>>

unwrap keys data with decrypt config.

fn annotation_id(&self) -> String

return the keywraper annotation id.

fn probe(&self, dc_param: &HashMap<String, Vec<Vec<u8>>>) -> bool

Check whether the driver could handle the decryption request.

Provided Methods

fn private_keys( &self, dc_param: &HashMap<String, Vec<Vec<u8>>> ) -> Option<Vec<Vec<u8>>>

private_keys (optional) gets the array of private keys. It is an optional implementation as in some key services, a private key may not be exportable (i.e. HSM) If not implemented, return None.

fn keyids_from_packet(&self, packet: String) -> Option<Vec<u64>>

keyids_from_packet (optional) gets a list of key IDs. This is optional as some encryption schemes may not have a notion of key IDs If not implemented, return None.

fn recipients(&self, recipients: String) -> Option<Vec<String>>

recipients (optional) gets a list of recipients. It is optional due to the validity of recipients in a particular encryption scheme If not implemented, return None.